SW-4501 Refactor the InventoryFilter into FilterMultiSelectContainer so we can easily use the filter multi select elsewhere
